requirements.txt
Faker==1.0.1
python-dateutil==2.7.5
PyYAML==3.13
selenium==3.141.0
six==1.12.0
text-unidecode==1.2
urllib3==1.24.1
